requirements.txt
numpy<2.0
pandas
matplotlib
seaborn
scikit-learn
tqdm
h5py
scipy
snntorch
tonic
wandb
nbformat
nbclient
nbconvert
